+++ cligen_custom.h
@@ -38,3 +38,7 @@ typedef struct {int a;} *cligen_handle;
 #else
 typedef void *cligen_handle; /* API */
 #endif
+
+/* Use a vector of help texts instead of a single string
+ */
+#define CO_HELPVEC