+++ cligen_expand.c
@@ -48,6 +48,7 @@
 #include "cligen_parsetree.h"
 #include "cligen_object.h"
 #include "cligen_handle.h"
+#include "cligen_match.h"
 #include "cligen_print.h"
 #include "cligen_expand.h"
 #include "cligen_syntax.h"
@@ -91,11 +92,17 @@ co_expand_sub(cg_obj  *co,
 	con->co_cvec = cvec_dup(co->co_cvec);
     if (co_callback_copy(co->co_callbacks, &con->co_callbacks) < 0)
 	return -1;
+#ifdef CO_HELPVEC
+    if (co->co_helpvec)
+	if ((con->co_helpvec = cvec_dup(co->co_helpvec)) == NULL)
+	    return -1;
+#else
     if (co->co_help)
 	if ((con->co_help = strdup(co->co_help)) == NULL){
 	    fprintf(stderr, "%s: strdup: %s\n", __FUNCTION__, strerror(errno));
 	    return -1;
 	}
+#endif
     if (co->co_type == CO_VARIABLE){
 	if (co->co_expand_fn_str)
 	    if ((con->co_expand_fn_str = strdup(co->co_expand_fn_str)) == NULL){
@@ -153,9 +160,19 @@ transform_var_to_cmd(cg_obj *co,
 	free(co->co_command);
     co->co_command = cmd; 
     if (helptext){
+#ifdef CO_HELPVEC
+	if (co->co_helpvec){
+	    cvec_free(co->co_helpvec);
+	    co->co_helpvec = NULL;
+	}
+	/* helpstr can be on the form "txt1\n    txt2" */
+	if (cligen_txt2cvv(helptext, &co->co_helpvec) < 0)
+	    return -1;
+#else
 	if (co->co_help)
 	    free(co->co_help);
 	co->co_help = helptext; 
+#endif
     }
     if (co->co_expandv_fn)
 	co->co_expandv_fn = NULL;
@@ -419,6 +436,12 @@ pt_expand_fnv(cligen_handle h,
 	/* 'escaped' always points to mutable string */
 	if (transform_var_to_cmd(con, (char*)escaped, helpstr) < 0)
 	    goto done;
+#ifdef CO_HELPVEC
+	if (helpstr){
+	    free(helpstr);
+	    helpstr = NULL;
+	}
+#endif
     }
     if (commands)
 	cvec_free(commands);